the about page. Added new rows
I added 2 new rows for the pictures of the interns  which are going to
be placed at. I had to also add a tag to center all the text at the
center. the 1st 2 Rows were oky. but when I added the last 2 the text
was aligned left. please have a look at that. Thankx