Updated the boi for Melusi and fixed the social network links
About page updated the content about Melusi Maphumulo's boi. Its not
complete will update it later.

style.css added new stlye.... taggeted the links on figcaption tag and
changed the color of the links since it was the same as the div which
slides up on hover together with the links. please check if it blands in
because I just changed the opacity